+++ mixin/index.test.js
@@ -1,9 +1,10 @@
-let Vuex = require('vuex')
-let { TestTime } = require('@logux/core')
-let { mount, createLocalVue } = require('@vue/test-utils')
-let { delay } = require('nanodelay')
+import { mount, createLocalVue } from '@vue/test-utils'
+import { TestTime } from '@logux/core'
+import { delay } from 'nanodelay'
+import { jest } from '@jest/globals'
+import Vuex from 'vuex'
 
-let { createLogux, loguxMixin } = require('..')
+import { createLogux, loguxMixin } from '../index.js'
 
 let localVue = createLocalVue()
 
@@ -198,7 +199,7 @@ it('does not resubscribe on non-relevant props changes', () => {
   })
 
   component.trigger('click', { id: 2 })
-  expect(resubscriptions).toEqual(0)
+  expect(resubscriptions).toBe(0)
 })
 
 it('reports about subscription end', async () => {
@@ -275,6 +276,7 @@ it('works on channels size changes 123', () => {
   localVue.nextTick(() => {
     component.trigger('click', { ids: [1, 2] })
   })
+  // eslint-disable-next-line no-console
   expect(console.error).not.toHaveBeenCalled()
 })